Bug 777292 - Make sure that the Components.results values are interpreted as unsigned integers; r=bholley
UltraBlame original commit: 30ab4e1d48745b68392b493fc98ed97a6df476a1

+++ js/xpconnect/src/XPCComponents.cpp
@@ -1403,7 +1403,10 @@ nsXPCComponents_Results::NewResolve(nsIXPConnectWrappedNative *wrapper,
         nsresult rv;
         while (nsXPCException::IterateNSResults(&rv, &rv_name, nullptr, &iter)) {
             if (!strcmp(name.ptr(), rv_name)) {
-                jsval val = JS_NumberValue((double)rv);
+                
+                
+                
+                jsval val = JS_NumberValue((double)(uint32_t)rv);
 
                 *objp = obj;
                 if (!JS_DefinePropertyById(cx, obj, id, val,
